Add files via upload
Made it so that you can input "today" in the second line of the inputs file and the code will use that as the end date for the search. This will save time of people needing to change that date to the current date every time they run the code.